Resuming blast activity after disk installations
The disk to storage the NR database was installed in tanto and now it’s
time to make available again the BLAST service. However, I have found
several problems probably derived of my strategy of refactoring the
styles, etc. As a first step, I have fixed the problem with the styles.
But some of the pages that should be able to watch are not working
properly. So I need to travel in time to check how the previous version
was working.